+++ services/car-service/src/main/java/be/ucll/da/carservice/domain/CarRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
 
-public interface CarRepository extends CrudRepository<Car, Long> {
+public interface CarRepository extends CrudRepository<Car, Integer> {
 
     List<Car> findAllByType(String type);
     List<Car> findAllByOwner(String owner);
